Working Game Commit
The game is now fully playable with an end to the game and a computer
player in a full GUI.

Things that are still not up to par:
The game board is displayed as a full single line, rather than 2
staggered lines. The focus was more on the get the game working first.